+++ packages/stargate/src/events.ts
@@ -1,5 +1,5 @@
 import { fromUtf8 } from "@cosmjs/encoding";
-import { tendermint34 } from "@cosmjs/tendermint-rpc";
+import { tendermint34, tendermint37 } from "@cosmjs/tendermint-rpc";
 
 /**
  * An event attribute.
@@ -30,16 +30,16 @@ export interface Event {
 }
 
 /**
- * Takes a Tendemrint 0.34 event with binary encoded key and value
+ * Takes a Tendermint 0.34 or 0.37 event with binary encoded key and value
  * and converts it into an `Event` with string attributes.
  */
-export function fromTendermint34Event(event: tendermint34.Event): Event {
+export function fromTendermintEvent(event: tendermint34.Event | tendermint37.Event): Event {
   return {
     type: event.type,
     attributes: event.attributes.map(
       (attr): Attribute => ({
-        key: fromUtf8(attr.key, true),
-        value: fromUtf8(attr.value, true),
+        key: typeof attr.key == "string" ? attr.key : fromUtf8(attr.key, true),
+        value: typeof attr.value == "string" ? attr.value : fromUtf8(attr.value, true),
       }),
     ),
   };