+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to
 - @cosmjs/proto-signing: Remove `fromJSON`/`toJSON` from `TsProtoGeneratedType`
   such that generated types are not required to generate those anymore. The
   methods were provided by ts-proto but we never needed them. ([#1329])
+- @cosmjs/stargate: Rename `fromTendermint34Event` to `fromTendermintEvent` and
+  let it support both Tendermint 0.34 and 0.37 events as input.
 
 [#1002]: https://github.com/cosmos/cosmjs/issues/1002
 [#1240]: https://github.com/cosmos/cosmjs/pull/1240